+++ lib/td_qx/data_views/queryable.ex
@@ -56,34 +56,8 @@ defmodule TdQx.DataViews.Queryable do
 
   def unfold([], resource_refs), do: resource_refs
 
-  def unfold(
-        %__MODULE__{type: "from", properties: properties} = queryable,
-        {resource_refs, queryables}
-      ) do
-    {resource_refs, from} = QueryableProperties.unfold(properties, queryable, resource_refs)
-    {resource_refs, queryables ++ [from]}
-  end
-
-  def unfold(%__MODULE__{type: "where", properties: properties}, {resource_refs, queryables}) do
-    where = QueryableProperties.unfold(properties)
-    {resource_refs, queryables ++ [where]}
-  end
-
-  def unfold(%__MODULE__{type: "select", properties: properties}, {resource_refs, queryables}) do
-    select = QueryableProperties.unfold(properties)
-    {resource_refs, queryables ++ [select]}
-  end
-
-  def unfold(
-        %__MODULE__{type: "join", properties: properties} = queryable,
-        {resource_refs, queryables}
-      ) do
-    {resource_refs, join} = QueryableProperties.unfold(properties, queryable, resource_refs)
-    {resource_refs, queryables ++ [join]}
-  end
-
-  def unfold(%__MODULE__{type: "group_by", properties: properties}, {resource_refs, queryables}) do
-    group_by = QueryableProperties.unfold(properties)
-    {resource_refs, queryables ++ [group_by]}
+  def unfold(queryable, {resource_refs, queryables}) do
+    {resource_refs, unfolded} = QueryableProperties.unfold(queryable, resource_refs)
+    {resource_refs, queryables ++ [unfolded]}
   end
 end