+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Fixed
+
+- [TD-6870] Added missing 'resource_ref' on GroupBy and Select queryables
+
 ## [6.12.0] 2024-09-23
 
 ### Added